% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumhmean.R
\name{cumhmean}
\alias{cumhmean}
\title{Cumulative Harmonic Mean}
\usage{
cumhmean(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A numeric vector of the same length as \code{x}. An \code{NA} value
   in \code{x} causes the corresponding and following elements of the return
   value to be \code{NA}.
}
\description{
Returns a vector whose elements are the cumulative \emph{harmonic mean} of
the elements of the argument.
}
\examples{
cumhmean(c(9, 1, 4, 0, 3, NA, 8, 5))

}
\author{
Arturo Erdely, \email{arturo.erdely@comunidad.unam.mx}
}
\references{
Kotz, S., Balakrishnan, N., Read, C.B, Vidakovic, B., Johnson, N.L. (2006)
   \emph{Encyclopedia of Statistical Sciences.} Wiley, New Jersey.
}
\seealso{
\code{\link{cumgmean}} for cummulative geometric mean, and
   \code{\link{cummean}} for cummulative mean.
}

