% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummode.R
\name{cummode}
\alias{cummode}
\title{Cumulative Mode}
\usage{
cummode(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A list of the same length as \code{x} with numeric vectors. \code{NA}
   values are also counted.
}
\description{
Returns a list whose elements are the cumulative \emph{statistical mode(s)}
of the elements of the argument.
}
\examples{
cummode(c(rep(1, 2), rep(12, 5), rep(44, 3), rep(8, 5), 55))

cummode(c(rep(1, 2), rep(12, 5), rep(44, 3), rep(8, 5), rep(NA, 7), 55))

cummode(runif(5))

cummode(c(rep("a", 2), rep("b", 5), rep("d", 3), rep("e", 5), rep(NA, 5)))

}
\author{
Arturo Erdely, \email{arturo.erdely@comunidad.unam.mx}
}
\references{
Kotz, S., Balakrishnan, N., Read, C.B, Vidakovic, B., Johnson, N.L. (2006)
   \emph{Encyclopedia of Statistical Sciences.} Wiley, New Jersey.
}
\seealso{
\code{\link{Mode}}
}

