% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumquant.R
\name{cumquant}
\alias{cumquant}
\title{Cumulative Quantile}
\usage{
cumquant(x, p, type = 7)
}
\arguments{
\item{x}{a numeric vector}

\item{p}{probability for the desired quantile}

\item{type}{See \code{quatile} in R base package}
}
\value{
A numeric vector of the same length as \code{x}. An \code{NA} value
   in \code{x} causes the corresponding and following elements of the return
   value to be \code{NA}.
}
\description{
Returns a vector whose elements are the cumulative \emph{quantile} of the
elements of the argument.
}
\examples{
y <- c(9, 1, 3, 0, NA, 2, 5)
cummedian(y)
cumquant(y, 0.5)

z <- cumquant(rcauchy(10000), 0.75)
head(z); tail(z)

}
\author{
Arturo Erdely, \email{arturo.erdely@comunidad.unam.mx}
}
\references{
Kotz, S., Balakrishnan, N., Read, C.B, Vidakovic, B., Johnson, N.L. (2006)
   \emph{Encyclopedia of Statistical Sciences.} Wiley, New Jersey.
}

