\name{cumstats-package}
\alias{cumstats-package}
\alias{cumstats}
\docType{package}
\title{
\packageTitle{cumstats}
}
\description{
\packageDescription{cumstats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cumstats}
\packageIndices{cumstats}
Cumulative descriptive statistics for (arithmetic, geometric, harmonic) mean, median, mode, variance, skewness and kurtosis.
}
\author{
\packageAuthor{cumstats}

Maintainer: \packageMaintainer{cumstats}
}
%\references{
%
%}
%~~ Optionally other standard keywords, one per line, from ~~
%~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%
%}
