% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A numeric value of skewness. Returns \code{NA} if \code{x} contains
   \code{NA} value(s), and \code{NaN} if \code{length(unique(x))==1} is
   \code{TRUE}.
}
\description{
This function computes skewness of given numeric data.
}
\examples{
skewness(c(9, 1, 3, 0))

}
\author{
Adapted by Arturo Erdely, \email{arturo.erdely@comunidad.unam.mx},
   from \code{moments} R package by Lukasz Komsta.
}
\references{
Komsta, L. and Novomestky, F. (2015). \emph{moments: Moments, cumulants,
skewness, kurtosis and related tests.} R package version 0.14.
\url{https://CRAN.R-project.org/package=moments}
}
\seealso{
\code{\link{cumskew}} for cummulative skewness.
}

